code_compiler ajout validation logique

+++ automation_code_builder/main.cpp
@@ -2,7 +2,7 @@
 #include "code_builder_factory.hpp"
 #include "validator.hpp"
 
-#define SCHEMA "../../schema.json"
+const string validation_schema_path = "../../schema.json";
 
 int main(int argc, char* argv[])
 {
@@ -21,21 +21,29 @@ int main(int argc, char* argv[])
     }
     json json = json::parse(file);
     file.close();
+    Automation_Validator validator(validation_schema_path);
+    auto validation_result = validator.validate(json);
     //1. TODO arthur: valider le schéma json
-    validJson(SCHEMA, json);
-    //2. TODO arthur: valider que les infos renseignées son supportées (target, devices type, références)
-    //3. TODO oussama: à compléter et tester. Générer le code qui décrit les settings 
-    filesystem::path output_directory ("generated_code");
-    cout << "Generating code to: " << filesystem::absolute(output_directory) << endl;
-    if(filesystem::exists(output_directory))
+    if(validation_result)
     {
-        filesystem::remove_all(output_directory);
-    }
-    filesystem::create_directory(output_directory);
+        //2. TODO arthur: valider que les infos renseignées son supportées (target, devices type, références)
+        //3. TODO oussama: à compléter et tester. Générer le code qui décrit les settings 
+        filesystem::path output_directory ("generated_code");
+        cout << "Generating code to: " << filesystem::absolute(output_directory) << endl;
+        if(filesystem::exists(output_directory))
+        {
+            filesystem::remove_all(output_directory);
+        }
+        filesystem::create_directory(output_directory);
 
-    ICodeBuilder* builder = CodeBuilderFactory::build(json["target"]);
-    builder->build(json, output_directory);
-    cout << "Completed" << endl;
+        ICodeBuilder* builder = CodeBuilderFactory::build(json["target"]);
+        builder->build(json, output_directory);
+        cout << "Completed" << endl;
+    }
+    else
+    {
+        cout << "'" << file_path << "' has an invalid schema : " << validation_result.message << endl;
+    }
 
     return 0;
 }
